containerd watcher resyncs on missed events better
Resync the whole state and restart listening for events at the resync
timestamp.

If we have missed events for a container, in particular the create
event, we are potentially out of sync. We now detect this and trigger a
state rebuild with attached handlers and event listeners. We rely on the
ignore map to represent containers we have seen (and want to ignore).
Containers are in the ignore map unless we have not seen them, or are
handling their start event.

Signed-off-by: Ray Bejjani <ray@covalent.io>